refactor: convert component to typescript
